+++ pyproject.toml
@@ -82,7 +82,7 @@ repository = "https://github.com/cambiegroup/flowchem"
 
 [project.scripts]
 flowchem = "flowchem.__main__:main"
-flowchem-autodiscover = "flowchem.autodiscover:main"
+flowchem-autodiscover = "flowchem.utils.autodiscover:main"
 
 [tool.setuptools]
 package-dir = {"" = "src"}
