+++ src/flowchem/devices/mettlertoledo/icir.py
@@ -12,8 +12,8 @@ from flowchem.components.analytics.ir import IRSpectrum
 from flowchem.devices.flowchem_device import DeviceInfo
 from flowchem.devices.flowchem_device import FlowchemDevice
 from flowchem.devices.mettlertoledo.icir_control import IcIRControl
-from flowchem.exceptions import DeviceError
-from flowchem.people import *
+from flowchem.utils.exceptions import DeviceError
+from flowchem.utils.people import *
 
 
 class ProbeInfo(BaseModel):
