+++ src/flowchem/devices/vacuubrand/cvc3000_finder.py
@@ -5,7 +5,7 @@ from textwrap import dedent
 from loguru import logger
 
 from flowchem.devices.vacuubrand.cvc3000 import CVC3000
-from flowchem.exceptions import InvalidConfiguration
+from flowchem.utils.exceptions import InvalidConfiguration
 
 
 # noinspection PyProtectedMember
