+++ src/flowchem/devices/vapourtec/r4_heater.py
@@ -14,8 +14,8 @@ from flowchem.components.technical.temperature import TempRange
 from flowchem.devices.flowchem_device import DeviceInfo
 from flowchem.devices.flowchem_device import FlowchemDevice
 from flowchem.devices.vapourtec.r4_heater_channel_control import R4HeaterChannelControl
-from flowchem.exceptions import InvalidConfiguration
-from flowchem.people import *
+from flowchem.utils.exceptions import InvalidConfiguration
+from flowchem.utils.people import *
 
 try:
     from flowchem_vapourtec import VapourtecR4Commands
