--- src/flowchem/server/sample_configuration.toml
@@ -1,30 +1,0 @@
-config_version = "1.0"
-simulation = true
-
-[device.donor]
-type = "Elite11InfuseOnly"
-port = "COM11"
-address = 0
-diameter = "4.6 mm"
-syringe_volume = "1 ml"
-
-[device.activator]
-type = "Elite11InfuseOnly"
-port = "COM11"
-address= 1
-diameter = "4.6 mm"
-syringe_volume = "1 ml"
-
-[device.quencher]
-type = "AxuraCompactPump"
-mac_address = "00:80:A3:BA:C3:4A"
-max_pressure = "13 bar"
-
-[device.sample-loop]
-type = "ViciValve"
-port = "COM13"
-address = 0
-
-[device.chiller]
-type = "HubeerChiller"
-port = "COM3"
