+++ src/flowchem/utils/autodiscover.py
